move uniform buffer fill method to its own module
