+++ crates/re_renderer/src/allocator/mod.rs
@@ -4,85 +4,9 @@
 //! follows some more complex strategy for efficient re-use and sub-allocation of wgpu resources.
 
 mod cpu_write_gpu_read_belt;
+mod uniform_buffer_fill;
 
 pub use cpu_write_gpu_read_belt::{CpuWriteGpuReadBelt, CpuWriteGpuReadBuffer};
-
-use crate::{wgpu_resources::BindGroupEntry, DebugLabel, RenderContext};
-
-struct UniformBufferAlignmentCheck<T> {
-    pub _marker: std::marker::PhantomData<T>,
-}
-impl<T> UniformBufferAlignmentCheck<T> {
-    // TODO: document why
-    const CHECK: () = assert!(
-        std::mem::align_of::<T>() >= 256,
-        "Uniform buffers need to be aligned to 256 bytes. Use `#[repr(C, align(256))]`"
-    );
-}
-/// Utility for fast & efficient creation of uniform buffers from a series of structs.
-///
-/// For subsequent frames, this will automatically not allocate any resources (thanks to our buffer pooling mechanism).
-///
-/// TODO(#1383): We could do this on a more complex stack allocator.
-pub fn create_and_fill_uniform_buffer_batch<T: bytemuck::Pod>(
-    ctx: &RenderContext,
-    label: DebugLabel,
-    content: impl ExactSizeIterator<Item = T>,
-) -> Vec<BindGroupEntry> {
-    #[allow(clippy::let_unit_value)]
-    let _ = UniformBufferAlignmentCheck::<T>::CHECK;
-
-    let num_buffers = content.len() as u64;
-    let element_size = std::mem::size_of::<T>() as u64;
-
-    assert!(
-        element_size > 0,
-        "Uniform buffer need to have a non-zero size"
-    );
-
-    let buffer = ctx.gpu_resources.buffers.alloc(
-        &ctx.device,
-        &crate::wgpu_resources::BufferDesc {
-            label,
-            size: num_buffers * element_size,
-            usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-            mapped_at_creation: false,
-        },
-    );
-
-    let mut staging_buffer = ctx.cpu_write_gpu_read_belt.lock().allocate::<T>(
-        &ctx.device,
-        &ctx.gpu_resources.buffers,
-        1,
-    );
-    staging_buffer.extend(content);
-    staging_buffer.copy_to_buffer(
-        ctx.active_frame
-            .frame_global_command_encoder
-            .lock()
-            .get_or_create(&ctx.device),
-        &buffer,
-        0,
-    );
-
-    (0..num_buffers)
-        .into_iter()
-        .map(|i| BindGroupEntry::Buffer {
-            handle: buffer.handle,
-            offset: i * element_size,
-            size: Some(std::num::NonZeroU64::new(element_size).unwrap()),
-        })
-        .collect()
-}
-
-/// See [`create_and_fill_uniform_buffer`].
-pub fn create_and_fill_uniform_buffer<T: bytemuck::Pod>(
-    ctx: &mut RenderContext,
-    label: DebugLabel,
-    content: T,
-) -> BindGroupEntry {
-    create_and_fill_uniform_buffer_batch(ctx, label, std::iter::once(content))
-        .into_iter()
-        .next()
-        .unwrap()
-}
+pub use uniform_buffer_fill::{
+    create_and_fill_uniform_buffer, create_and_fill_uniform_buffer_batch,
+};
